feat:add ask-user tool and modularize react tools
