+++ packages/dbgpt-app/src/dbgpt_app/openapi/api_v1/tools/execute_analysis.py
@@ -1,0 +1,115 @@
+"""execute_analysis tool — quick Excel/CSV analysis via code server."""
+
+import json
+from typing import Any, Dict, List
+
+from dbgpt.agent.resource.tool.base import tool
+
+
+def make_execute_analysis(react_state: Dict[str, Any]):
+    @tool(description="Execute quick analysis on uploaded Excel/CSV file.")
+    async def execute_analysis() -> str:
+        from dbgpt._private.config import Config
+        from dbgpt_app.openapi.api_v1.agentic_data_api import get_code_server
+
+        CFG = Config()
+
+        def _is_excel_skill(meta) -> bool:
+            name = (meta.name or "").lower()
+            desc = (meta.description or "").lower()
+            tags = [tag.lower() for tag in (meta.tags or [])]
+            return any(
+                token in name or token in desc or token in tags
+                for token in ["excel", "xlsx", "xls", "spreadsheet"]
+            )
+
+        matched = react_state.get("matched")
+        if not react_state.get("file_path"):
+            return json.dumps(
+                {"chunks": [{"output_type": "text", "content": "No file to analyze"}]},
+                ensure_ascii=False,
+            )
+        if matched and not _is_excel_skill(matched.metadata):
+            return json.dumps(
+                {
+                    "chunks": [
+                        {
+                            "output_type": "text",
+                            "content": "Selected skill is not for Excel analysis",
+                        }
+                    ]
+                },
+                ensure_ascii=False,
+            )
+        code_server = await get_code_server(CFG.SYSTEM_APP)
+        analysis_code = """
+import json
+import pandas as pd
+
+file_path = r"{file_path}"
+if file_path.lower().endswith((".xls", ".xlsx")):
+    df = pd.read_excel(file_path)
+else:
+    df = pd.read_csv(file_path)
+summary = {{
+    "shape": list(df.shape),
+    "columns": list(df.columns),
+    "dtypes": {{col: str(dtype) for col, dtype in df.dtypes.items()}},
+    "head": df.head(5).to_dict(orient="records"),
+}}
+print(json.dumps(summary, ensure_ascii=False))
+""".format(file_path=react_state["file_path"])
+        result = await code_server.exec(analysis_code, "python")
+        output_text = (
+            result.output.decode("utf-8") if isinstance(result.output, bytes) else ""
+        )
+        chunks: List[Dict[str, Any]] = [
+            {"output_type": "code", "content": analysis_code.strip()}
+        ]
+        if output_text:
+            try:
+                summary = json.loads(output_text)
+                chunks.append({"output_type": "json", "content": summary})
+                head_rows = summary.get("head")
+                columns = summary.get("columns")
+                if isinstance(head_rows, list) and isinstance(columns, list):
+                    chunks.append(
+                        {
+                            "output_type": "table",
+                            "content": {
+                                "columns": [
+                                    {"title": col, "dataIndex": col, "key": col}
+                                    for col in columns
+                                ],
+                                "rows": head_rows,
+                            },
+                        }
+                    )
+                numeric_columns = [
+                    col
+                    for col, dtype in (summary.get("dtypes") or {}).items()
+                    if "int" in dtype or "float" in dtype
+                ]
+                if numeric_columns and isinstance(head_rows, list):
+                    series_col = numeric_columns[0]
+                    data = [
+                        {"x": idx + 1, "y": row.get(series_col)}
+                        for idx, row in enumerate(head_rows)
+                        if row.get(series_col) is not None
+                    ]
+                    if data:
+                        chunks.append(
+                            {
+                                "output_type": "chart",
+                                "content": {
+                                    "data": data,
+                                    "xField": "x",
+                                    "yField": "y",
+                                },
+                            }
+                        )
+            except Exception:
+                chunks.append({"output_type": "text", "content": output_text})
+        return json.dumps({"chunks": chunks}, ensure_ascii=False)
+
+    return execute_analysis
